# src/backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "T-Coffee API"
  },
  "host": "192.168.0.76:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Get a product by ID",
        "description": "",
        "operationId": "getProductById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Invalid ID supplied"
          },
          "400": {
            "description": "Please provide product id"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateProduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be updated",
            "schema": {
              "$ref": "#/definitions/updateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Delete a product",
        "description": "",
        "operationId": "deleteProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be deleted",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/order": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get all orders",
        "description": "Returns all orders from the system that the user has access to",
        "operationId": "getOrders",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create a new order",
        "description": "Create a new order",
        "operationId": "createOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Order object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/OrderCreateResponse"
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/order/{orderId}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get a order by ID",
        "description": "",
        "operationId": "getOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "500": {
            "description": "Invalid ID supplied"
          },
          "400": {
            "description": "Please provide order id"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Update an existing order",
        "description": "",
        "operationId": "updateOrder",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Order object that needs to be updated",
            "schema": {
              "$ref": "#/definitions/OrderCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Order updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Delete an order",
        "description": "",
        "operationId": "deleteOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be deleted",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Order deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "variants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Variant"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      }
    },
    "Variant": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Option"
          }
        }
      }
    },
    "Option": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "additional_price": {
          "type": "number",
          "format": "double"
        },
        "variant_id": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        }
      }
    },
...
    "updateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Cappuccino",
          "required": [
            "false"
          ]
        },
        "price": {
          "type": "number",
          "format": "double",
          "example": 39000,
          "required": [
            "false"
          ]
        },
        "description": {
          "type": "string",
          "example": "Cappuccino is a coffee drink that today is typically composed of double espresso and hot milk, with the surface topped with foamed milk.",
          "required": [
            "false"
          ]
        },
        "image": {
          "type": "string",
          "example": "https://www.starbucks.com.vn/media/catalog/product/cache/1/image/9df78eab33525d08d6e5fb8d27136e95/c/a/cappuccino.jpg",
          "required": [
            "false"
          ]
        },
        "variantId": {
          "type": "array",
          "required": [
            "false"
          ],
          "items": {
            "type": "string"
          },
          "example": [
            "size",
            "sugar",
            "topping"
          ]
        },
        "categoryId": {
          "type": "array",
          "required": [
            "false"
          ],
          "items": {
            "type": "string"
          },
          "example": [
            "coffee",
            "drinks"
          ]
        }
      }
    },
    "OrderCreate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "481034707442110021390829492_1724006525222"
        },
        "customer_id": {
          "type": "string",
          "example": "759280903754672026"
        },
        "shipping_address": {
          "type": "string",
          "example": "123 Nguyen Van Linh, Da Nang"
        },
        "total_amount": {
          "type": "number",
          "example": 39000
        },
        "total_products": {
          "type": "number",
          "example": 2
        },
        "shipping_fee": {
          "type": "number",
          "example": 10000
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "quantity": {
                "type": "number",
                "example": 2
              }
            }
          }
        }
      }
    },
    "OrderCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "481034707442110021390829492_1724006525222"
        },
        "shipping_address": {
          "type": "string",
          "example": "123 Nguyen Van Linh, Da Nang"
        },
        "total_amount": {
          "type": "number",
          "example": 39000
        },
        "shipping_fee": {
          "type": "number",
          "example": 10000
        },
        "customer_id": {
          "type": "string",
          "example": "759280903754672026"
        },
        "order_date": {
          "type": "string",
          "example": "2024-08-26T04:09:50.685Z"
        },
        "updatedAt": {
          "type": "string",
          "example": "2024-08-26T04:09:50.685Z"
        },
        "createdAt": {
          "type": "string",
          "example": "2024-08-26T04:09:50.685Z"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "481034707442110021390829492_1724006525222"
        },
        "shipping_address": {
          "type": "string",
          "example": "123 Nguyen Van Linh, Da Nang"
        },
        "total_amount": {
          "type": "number",
          "example": 39000
        },
        "shipping_fee": {
          "type": "number",
          "example": 10000
        },
        "order_date": {
          "type": "string",
          "example": "2024-08-26T04:09:50.685Z"
        },
        "customer": {
          "type": "object",
          "$ref": "#/definitions/Customer"
        },
        "discounts": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "Discount 10%"
              },
              "description": {
                "type": "string",
                "example": "Discount 10% for all products"
              },
              "discount_percent": {
                "type": "number",
                "example": 10
              }
            }
          }
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "Caramel Macchiato"
              },
              "description": {
                "type": "string",
                "example": "Caramel Macchiato is a coffee drink that is made with espresso, steamed milk, and caramel syrup."
              },
              "price": {
                "type": "number",
                "example": 39000
              },
              "image": {
                "type": "string",
                "example": "https://www.starbucks.com.vn/media/catalog/product/cache/1/image/9df78eab33525d08d6e5fb8d27136e95/c/a/caramel-macchiato.jpg"
              },
              "quantity": {
                "type": "number",
                "example": 2
              }
            }
          }
        }
      }
    },
    "Customer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "759280903754672026"
        },
        "name": {
          "type": "string",
          "example": "Nguyen Van A"
        },
        "phone": {
          "type": "string",
          "example": "0123456789"
        }
      }
    }
  }
}